+++ ui/memory.go
@@ -235,9 +235,9 @@ func fetchMemoryDbAlloc(ctx context.Context, dballocChannel chan<- [2]string, lc
 	for {
 		select {
 		case <-ticker.C:
-			lookup = GlobalQueryMap[Instances[CurrConn].DBMS]
-			global_alloc = queries.GetLongQuery(Instances[CurrConn].Driver, lookup["global_alloc"]())
-			alloc_by_area := queries.GetLongQuery(Instances[CurrConn].Driver, lookup["spec_alloc"]())
+			lookup = GlobalQueryMap[Instances[ActiveConns[0]].DBMS]
+			global_alloc = queries.GetLongQuery(Instances[ActiveConns[0]].Driver, lookup["global_alloc"]())
+			alloc_by_area := queries.GetLongQuery(Instances[ActiveConns[0]].Driver, lookup["spec_alloc"]())
 
 			messages[0] += "\n\n   Total allocated\n\n"
 			messages[1] += "\n\n" + global_alloc[0][0] + "\n\n"
@@ -322,8 +322,8 @@ func fetchMemoryUserAlloc(ctx context.Context, usrallocChannel chan<- [2]string,
 	for {
 		select {
 		case <-ticker.C:
-			lookup = GlobalQueryMap[Instances[CurrConn].DBMS]
-			usr_alloc = queries.GetLongQuery(Instances[CurrConn].Driver, lookup["user_alloc"]())
+			lookup = GlobalQueryMap[Instances[ActiveConns[0]].DBMS]
+			usr_alloc = queries.GetLongQuery(Instances[ActiveConns[0]].Driver, lookup["user_alloc"]())
 
 			messages[0] += "\n\n   User\n\n"
 			messages[1] += fmt.Sprintf("\n\n%-11v %-9v\n", "Current", "(Max)")
@@ -383,8 +383,8 @@ func fetchMemoryHardwAlloc(ctx context.Context, hardwallocChannel chan<- [2]stri
 	for {
 		select {
 		case <-ticker.C:
-			lookup = GlobalQueryMap[Instances[CurrConn].DBMS]
-			ramndisk_alloc = queries.GetLongQuery(Instances[CurrConn].Driver, lookup["ramdisk_alloc"]())
+			lookup = GlobalQueryMap[Instances[ActiveConns[0]].DBMS]
+			ramndisk_alloc = queries.GetLongQuery(Instances[ActiveConns[0]].Driver, lookup["ramdisk_alloc"]())
 
 			messages[0] += "\n\n\n\n                   Disk\n                    RAM"
 			messages[1] += fmt.Sprintf("\n\n%-11v %-9v\n", "Current", "(Max)")