+++ ui/transactions.go
@@ -73,10 +73,27 @@ func DisplayTransactions() {
 		LastInputTime = time.Now()
 
 		switch k.Key {
-		case keyboard.KeyArrowLeft:
-			CurrRotateLeft()
-		case keyboard.KeyArrowRight:
-			CurrRotateRight()
+		case 'p', 'P':
+			State = types.PROCESSLIST
+			cancel()
+		case 'd', 'D':
+			State = types.DB_DASHBOARD
+			cancel()
+		case 'm', 'M':
+			State = types.MEM_DASHBOARD
+			cancel()
+		case 'e', 'E':
+			State = types.ERR_LOG
+			cancel()
+		case 'l', 'L':
+			State = types.LOCK_LOG
+			cancel()
+		case 'r', 'R':
+			State = types.REPLICATION
+			cancel()
+		case 'c', 'C':
+			State = types.CONFIGS
+			cancel()
 		case '?':
 			State = types.MENU
 			cancel()
@@ -85,12 +102,9 @@ func DisplayTransactions() {
 		case keyboard.KeyEsc:
 			State = Laststate
 			cancel()
-		case '+':
-			Interval += 100 * time.Millisecond
-		case '-':
-			if Interval > 100*time.Millisecond {
-				Interval -= 100 * time.Millisecond
-			}
+		case 'q', 'Q':
+			State = types.QUIT
+			cancel()
 		}
 	}
 
@@ -138,10 +152,11 @@ func fetchTransactions(ctx context.Context,
 	for {
 		select {
 		case <-ticker.C:
-			lookup = GlobalQueryMap[Instances[CurrConn].DBMS]
-			messages = queries.GetLongQuery(Instances[CurrConn].Driver, lookup["transactions"]())
+			lookup = GlobalQueryMap[Instances[ActiveConns[0]].DBMS]
+			messages = queries.GetLongQuery(Instances[ActiveConns[0]].Driver, lookup["transactions"]())
 
 			txnsChannel <- messages
+			messages = [][]string{}
 
 		case <-ctx.Done():
 			return
@@ -184,7 +199,11 @@ func writeTransactions(ctx context.Context,
 					color = text.WriteCellOpts(cell.FgColor(cell.ColorWhite))
 				}
 				colorflipper *= -1
-				txns_text.Write(fmt.Sprintf("%-5v %-25v %-15v %-15v %-90v\n", line[0], line[1], line[2], line[3], line[4]), color)
+				out := fmt.Sprintf("%-5v %-25v %-15v %-15v %-90v\n", line[0], line[1], line[2], line[3], line[4])
+				if len(out) > 128 {
+					out = out[:128]
+				}
+				txns_text.Write(out, color)
 			}
 
 		case <-ctx.Done():